renamed sample app package name
