+++ samples/kotlin/src/main/java/com/forgerock/kotlinapp/NodeDialogFragment.kt
@@ -1,4 +1,4 @@
-package com.forgerock.androidapp2
+package com.forgerock.kotlinapp
 
 import android.content.Context
 import android.os.Bundle
@@ -10,16 +10,21 @@ import android.widget.Button
 import android.widget.RadioButton
 import androidx.appcompat.widget.AppCompatTextView
 import androidx.fragment.app.DialogFragment
+import com.bingo.foo.MainActivity
+import com.google.android.material.textfield.TextInputEditText
+import com.google.android.material.textfield.TextInputLayout
 import org.forgerock.android.auth.Node
 import org.forgerock.android.auth.callback.ChoiceCallback
+import org.forgerock.android.auth.callback.NameCallback
+import org.forgerock.android.auth.callback.PasswordCallback
 
 
-class ChoiceDialogFragment: DialogFragment() {
+class NodeDialogFragment: DialogFragment() {
     private var listener: MainActivity? = null
     private var node: Node? = null
     companion object {
-        fun newInstance(node: Node?): ChoiceDialogFragment {
-            return ChoiceDialogFragment().apply {
+        fun newInstance(node: Node?): NodeDialogFragment {
+            return NodeDialogFragment().apply {
                 arguments = Bundle().apply {
                     this.putSerializable("NODE", node)
                 }
@@ -28,17 +33,23 @@ class ChoiceDialogFragment: DialogFragment() {
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? = inflater.inflate(R.layout.choice_node, container, false)
+                              savedInstanceState: Bundle?): View? = inflater.inflate(R.layout.fragment_node, container, false)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         node = arguments?.getSerializable("NODE") as Node?
+        val choiceNode: View = view.findViewById(R.id.choice_node)
         val choiceCallBack = node?.getCallback(ChoiceCallback::class.java)
         choiceCallBack?.let {
-            val yesRadioButton: RadioButton = view.findViewById(R.id.yes_radio_button)
-            val title: AppCompatTextView = view.findViewById(R.id.title)
-            title.text = node?.getCallback(ChoiceCallback::class.java)?.prompt
-            val noRadioButton: RadioButton = view.findViewById(R.id.no_radio_button)
+            val yesRadioButton: RadioButton = choiceNode.findViewById(R.id.yes_radio_button)
+            val noRadioButton: RadioButton = choiceNode.findViewById(R.id.no_radio_button)
+            noRadioButton.text = choiceCallBack.choices[1]
+            choiceCallBack.choices.takeIf { it1 -> it1.count() > 1 }.apply {
+                yesRadioButton.text = choiceCallBack.choices[0] ?: "Yes"
+                noRadioButton.text = choiceCallBack.choices[1] ?: "No"
+            }
+            val title: AppCompatTextView = choiceNode.findViewById(R.id.title)
+            title.text = choiceCallBack.prompt
             yesRadioButton.isChecked = it.defaultChoice == 0
             noRadioButton.isChecked = it.defaultChoice == 1
             yesRadioButton.setOnClickListener {
@@ -50,9 +61,26 @@ class ChoiceDialogFragment: DialogFragment() {
                 node?.getCallback(ChoiceCallback::class.java)?.setSelectedIndex(1)
             }
         }
+        node?.callbacks?.forEach {
+            when (it.type) {
+                "NameCallback" -> {
+                    (view.findViewById(R.id.usernameLayout) as? TextInputLayout)?.visibility = View.VISIBLE
+                }
+                "PasswordCallback" -> {
+                    (view.findViewById(R.id.passwordLayout) as? TextInputLayout)?.visibility = View.VISIBLE
+                }
+                "ChoiceCallback" -> {
+                    choiceNode.visibility = View.VISIBLE
+                }
+            }
+        }
+        val username: TextInputEditText = view.findViewById(R.id.username)
+        val password: TextInputEditText = view.findViewById(R.id.password)
         val next: Button = view.findViewById(R.id.next)
         next.setOnClickListener {
             dismiss()
+            node?.getCallback(NameCallback::class.java)?.setName(username.text.toString())
+            node?.getCallback(PasswordCallback::class.java)?.setPassword(password.text.toString().toCharArray())
             node?.next(context, listener)
         }
         val cancel: Button = view.findViewById(R.id.cancel)
